SandboxSimulation done
SandboxSimulation itself is done, will do SimulationNode next.

The problem is the react-contextmenu plugin, it creates a DOM element for the trigger area and that fucks-up with react-draggable. I will replace react-contextmenu with a plugin that does not create its own DOM element for the trigger area.
cleanup: remove global layout debug style